Improve test coverage for commands
