+++ tests/ballerina-tools-integration-test/src/test/resources/testng.xml
@@ -26,9 +26,13 @@
     <test name="ballerina-packaging-sample-tests" preserve-order="true" parallel="false">
         <classes>
             <class name="org.ballerinalang.test.packaging.PackagingTestCase"/>
+            <class name="org.ballerinalang.test.packaging.PackagingInitTestCase"/>
+            <class name="org.ballerinalang.test.packaging.PackagingTestExecutionTestCase"/>
+            <class name="org.ballerinalang.test.packaging.PackagingDocGenTestCase"/>
+            <class name="org.ballerinalang.test.packaging.PackagingListDependencyTestCase"/>
             <class name="org.ballerinalang.test.packaging.PackagingNegativeTestCase"/>
-            <class name="org.ballerinalang.test.grpc.ServicePackagingTestCase"/>
             <class name="org.ballerinalang.test.packaging.PackagingSingleBalBuildTestCase"/>
+            <class name="org.ballerinalang.test.grpc.ServicePackagingTestCase"/>
         </classes>
     </test>
